requirements.txt
asyncio==3.4.3
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.1
idna==3.4
kafka-python==2.0.2
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
tzdata==2023.3
urllib3==1.26.15
zope.interface==6.0
